Move the operation_configuration class to its own file
